refactor: move ProposalActionGroupConfig into models folder under respective module
