+++ src/proposals/proposal-actions/proposal-action-group-configs/proposal-action-group-configs.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ProposalActionGroupConfig } from '../models/proposal-action-group-config.model';
+import { ProposalActionGroupConfig } from './models/proposal-action-group-config.model';
 import { ProposalActionGroupConfigsResolver } from './proposal-action-group-configs.resolver';
 import { ProposalActionGroupConfigsService } from './proposal-action-group-configs.service';
